Paginate and extract image tag when using describe_images at ECR module (#1480)
### Summary
Follow up from
https://github.com/cartography-cncf/cartography/pull/1471/.

[describe_images](https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/ecr/client/describe_images.html)
returns `imageTags` as an array of tags, compared to the single
`imageTag`returned by
[list_images](https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/ecr/client/list_images.html).
Added as well more data to include a couple of the new fields to the
integration test.


### Related issues or links
- https://github.com/cartography-cncf/cartography/pull/1471


### Checklist

Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [x] Update/add unit or integration tests.
- [ ] Include a screenshot showing what the graph looked like before and
after your changes.
- [ ] Include console log trace showing what happened before and after
your changes.

If you are changing a node or relationship:
- [ ] Update the
[schema](https://github.com/lyft/cartography/tree/master/docs/root/modules)
and
[readme](https://github.com/lyft/cartography/blob/master/docs/schema/README.md).

If you are implementing a new intel module:
- [ ] Use the NodeSchema [data
model](https://cartography-cncf.github.io/cartography/dev/writing-intel-modules.html#defining-a-node).

---------

Signed-off-by: Hector Eryx Paredes Camacho <heryxpc@users.noreply.github.com>
Signed-off-by: Eryx Paredes <eryxp@lyft.com>
Co-authored-by: i_virus <chandan.chowdhury@hotmail.com>