Fix #1481: Separate EC2 template version fetching from template fetching (#1486)
### Summary
> Describe your changes.

This change separates the retrieval of launch templates and their
versions into distinct functions, resolving an issue where the
`aws_handle_regions` decorator returns an empty list instead of a tuple,
leading to a `ValueError` when unpacking.

I see that #845 specifically reverted a change that made
`aws_handle_regions` polymorphic in return type.


### Related issues or links
> Include links to relevant issues or other pages.

- https://github.com/lyft/cartography/issues/...


### Checklist

Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [ ] Update/add unit or integration tests.
- [ ] Include a screenshot showing what the graph looked like before and
after your changes.
- [ ] Include console log trace showing what happened before and after
your changes.

If you are changing a node or relationship:
- [ ] Update the
[schema](https://github.com/lyft/cartography/tree/master/docs/root/modules)
and
[readme](https://github.com/lyft/cartography/blob/master/docs/schema/README.md).

If you are implementing a new intel module:
- [ ] Use the NodeSchema [data
model](https://cartography-cncf.github.io/cartography/dev/writing-intel-modules.html#defining-a-node).

---------

Signed-off-by: Ishaan Verma <ishaanverma97@gmail.com>