Fix #1496: support neo4j driver version 5 (#1497)
### Summary
> Describe your changes.

If we update cartography from neo4j python driver 4 to 5, cartography
integration tests fail.

This is because Neo4j 5 is stricter about calling `consume()` on a
closed transaction. Currently, GraphStatement functions
`_run_iterative()` and `_run_noniterative()` all call `consume()`
incorrectly.

This PR straightens this out so that `consume()` is never called on a
closed transaction.


### Related issues or links
> Include links to relevant issues or other pages.

- https://github.com/lyft/cartography/issues/1496


### Checklist

Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [x] Update/add unit or integration tests.
- [ ] Include a screenshot showing what the graph looked like before and
after your changes.
- [x] Include console log trace showing what happened before and after
your changes.


### Before
![Screenshot 2025-03-26 at 11 32
12 PM](https://github.com/user-attachments/assets/1557a736-583c-4ad6-bddf-d96f59e24fa2)


### After
![Screenshot 2025-03-26 at 11 57
59 PM](https://github.com/user-attachments/assets/b650419a-7815-40df-8312-f512be238d15)



If you are changing a node or relationship:
- [ ] Update the
[schema](https://github.com/lyft/cartography/tree/master/docs/root/modules)
and
[readme](https://github.com/lyft/cartography/blob/master/docs/schema/README.md).

If you are implementing a new intel module:
- [ ] Use the NodeSchema [data
model](https://cartography-cncf.github.io/cartography/dev/writing-intel-modules.html#defining-a-node).

Signed-off-by: Alex Chantavy <alex@subimage.io>