feat: Support one-to-many in data model, implement iaminstanceprofiles (#1502)
### Summary
> Describe your changes.

Addresses discussion in #1465 and in this [slack
thread](https://cloud-native.slack.com/archives/C080M2LRLDA/p1743350053575729)
by adding support for one-to-many relationships in the Cartography data
model.

Puts it to use to fix #1365 by adding correct support for
iaminstanceprofiles. We now properly represent that AWS roles and EC2
instances have a many-to-many association.


### Related issues or links
> Include links to relevant issues or other pages.

- https://github.com/lyft/cartography/issues/1365


### Checklist

Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [x] Update/add unit or integration tests.
- [x] Include a screenshot showing what the graph looked like before and
after your changes.

![image](https://github.com/user-attachments/assets/2ccb7dc8-8320-42b3-b0cf-da81e6a016c6)

- [ ] Include console log trace showing what happened before and after
your changes.

If you are changing a node or relationship:
- [x] Update the
[schema](https://github.com/lyft/cartography/tree/master/docs/root/modules)
and
[readme](https://github.com/lyft/cartography/blob/master/docs/schema/README.md).

If you are implementing a new intel module:
- [x] Use the NodeSchema [data
model](https://cartography-cncf.github.io/cartography/dev/writing-intel-modules.html#defining-a-node).


This PR looks bigger than it actually is because of test data and tests.

---------

Signed-off-by: Alex Chantavy <alex@subimage.io>