Allow only specific aws regions to be synced with --aws-regions (#1542)
### Summary
> Describe your changes.

Adds `--aws-regions` CLI switch to support syncing of only specific
regions as an experimental feature.

At this time, this feature only supports syncing the same regions for
all accounts. Specifying this on a per-account basis is not currently
supported.

Also see
https://cloud-native.slack.com/archives/C080M2LRLDA/p1745250103100609
and
https://docs.google.com/document/d/1VyRKmB0dpX185I15BmNJZpfAJ_Ooobwz0U1WIhjDxvw/edit?tab=t.0#heading=h.orecc9yg8l30.

### Checklist

Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [x] Update/add unit or integration tests.
- [x] Include a screenshot showing what the graph looked like before and
after your changes.
![Screenshot 2025-04-26 at 4 25
54 PM](https://github.com/user-attachments/assets/1430e23f-41d2-46b8-9d08-122c935884cc)



- [ ] Include console log trace showing what happened before and after
your changes.

If you are changing a node or relationship:
- [ ] Update the
[schema](https://github.com/lyft/cartography/tree/master/docs/root/modules)
and
[readme](https://github.com/lyft/cartography/blob/master/docs/schema/README.md).

If you are implementing a new intel module:
- [ ] Use the NodeSchema [data
model](https://cartography-cncf.github.io/cartography/dev/writing-intel-modules.html#defining-a-node).

---------

Signed-off-by: Alex Chantavy <alex@subimage.io>