fix: display data on Intel CVE error (#1558)
### Summary
Small PR - fixes the formatting to display the data in Intel CVE error.

### Related issues or links


### Checklist

Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [ ] Update/add unit or integration tests.
- [ ] Include a screenshot showing what the graph looked like before and
after your changes.
- [ ] Include console log trace showing what happened before and after
your changes.

If you are changing a node or relationship:
- [ ] Update the
[schema](https://github.com/cartography-cncf/cartography/tree/master/docs/root/modules)
and
[readme](https://github.com/cartography-cncf/cartography/blob/master/docs/schema/README.md).

If you are implementing a new intel module:
- [ ] Use the NodeSchema [data
model](https://cartography-cncf.github.io/cartography/dev/writing-intel-modules.html#defining-a-node).

Signed-off-by: Emmanuel Ferdman <emmanuelferdman@gmail.com>