Add EfsMountTarget resource to AWS Module (#1573)
### Summary
This PR adds the AWS Efs MountTarget as a supported resource for the
AWS Module.


### Related issues or links
[Issue](https://github.com/cartography-cncf/cartography/issues/1552)


### Checklist

Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [x] Update/add unit or integration tests.
- [ ] Include a screenshot showing what the graph looked like before and
after your changes.
- [ ] Include console log trace showing what happened before and after
your changes.

If you are changing a node or relationship:
- [x] Update the
[schema](https://github.com/cartography-cncf/cartography/tree/master/docs/root/modules)
and
[readme](https://github.com/cartography-cncf/cartography/blob/master/docs/schema/README.md).

If you are implementing a new intel module:
- [x] Use the NodeSchema [data
model](https://cartography-cncf.github.io/cartography/dev/writing-intel-modules.html#defining-a-node).

Signed-off-by: d-aggarwal <drishtiaggarwal6284@gmail.com>