feat: Add SSMParameter resource to AWS module (#1576)
### Summary
> Describe your changes.
This commit adds support for ingesting AWS SSM Parameters as a supported
resource for the AWS Module.


### Related issues or links
(#1552)


### Checklist

Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [x] Update/add unit or integration tests.
- [ ] Include a screenshot showing what the graph looked like before and
after your changes.
- [ ] Include console log trace showing what happened before and after
your changes.

If you are changing a node or relationship:
- [x] Update the
[schema](https://github.com/cartography-cncf/cartography/tree/master/docs/root/modules)
and
[readme](https://github.com/cartography-cncf/cartography/blob/master/docs/schema/README.md).

If you are implementing a new intel module:
- [x] Use the NodeSchema [data
model](https://cartography-cncf.github.io/cartography/dev/writing-intel-modules.html#defining-a-node).

---------

Signed-off-by: Khush Agrawal <khushmagrawal15@gmail.com>
Co-authored-by: Jeremy Chapeau <113923302+jychp@users.noreply.github.com>