Ingest SentinelOne account nodes  (#1661)
### Summary

First PR for the SentinelOne intelmodule. This PR covers the ingestion
of the S1 account nodes which will serve as root nodes for the resources
of the assets and CVEs. It is possible to provide a set of account IDs
to ingest otherwise the module will ingest every account accessible to
the user that owns the API token.

I'll update the intelmodule and the corresponding schema docs once all
the intended resources are implemented.

Logs from running this locally:

![Screenshot 2025-07-01 at 9 55 23 a
m](https://github.com/user-attachments/assets/14c08b93-431d-49f9-bece-0ccba3c5416b)

API token was placed in the env variable `SENTINELONE_API_TOKEN`

### Related issues or links
- https://github.com/cartography-cncf/cartography/issues/1660

### Checklist

Provide proof that this works (this makes reviews move faster). Please
perform one or more of the following:
- [x] Update/add unit or integration tests.
- [ ] Include a screenshot showing what the graph looked like before and
after your changes.
- [x] Include console log trace showing what happened before and after
your changes.

If you are changing a node or relationship:
- [ ] Update the
[schema](https://github.com/cartography-cncf/cartography/tree/master/docs/root/modules)
and
[readme](https://github.com/cartography-cncf/cartography/blob/master/docs/schema/README.md).

If you are implementing a new intel module:
- [x] Use the NodeSchema [data
model](https://cartography-cncf.github.io/cartography/dev/writing-intel-modules.html#defining-a-node).

---------

Signed-off-by: Sergio Franco <sergiof@lyft.com>